% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stations.R
\name{cluster_stations}
\alias{cluster_stations}
\title{Cluster stations that are near one another.}
\usage{
cluster_stations(lon, lat, elev, dist_adj, elev_adj, h)
}
\arguments{
\item{lon}{A numeric vector of station longitudes.}

\item{lat}{A numeric vector of station latitudes.}

\item{elev}{A numeric vector of station elevations.}

\item{dist_adj}{Constant value in miles that scales the dissimilarity score
for geographic distance.}

\item{elev_adj}{Constant value that scales elevation dissimilarity.}

\item{h}{Maximum similarity score for the cluster analysis.}
}
\value{
A numeric vector of clusters, where each unique number represents a
cluster.
}
\description{
Uses complete linkage clustering where each cluster has a maximum similarity
score of h between furthest stations in a cluster. dist_adj can be used to
scale distances between stations and elev_adj can be used to scale elevation
between stations. For example, distance can be scaled so that dist_adj = 4
miles is a one unit similarity between stations and elevation can be scaled
so that elev_adj = 50 is a one unit similarity between stations. Then two
stations that are separated by 4 miles and have a difference of 50 units of
elevation have a total similarity score of 2. h is used to specify a maximum
similarity score between stations in a cluster.
}
\examples{
# Simple example
subset <- snowload2::ghcnd_stations[10500:10600,]
cluster_stations(subset$LONGITUDE, subset$LATITUDE, subset$ELEVATION,
                 dist_adj = 4, elev_adj = 50, h = 2)

# Practical use example
library(dplyr)
cluster_test <- snowload2::ghcnd_stations \%>\%
  filter(STATE \%in\% c("UT", "NV")) \%>\%
  group_by(STATE) \%>\%
  mutate(CLUST = cluster_stations(LONGITUDE, LATITUDE, ELEVATION,
                                  dist_adj = 4,
                                  elev_adj = 50,
                                  h = 2)) \%>\%
  ungroup() \%>\%
  mutate(CLUST = sprintf(paste0(STATE, "\%04d"), CLUST))


}
