% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_maxes}
\alias{county_maxes}
\title{Max 1,2,...,10 day snowfall totals for each county.}
\format{
A data frame with 31430 rows and 4 variables:

 \describe{
 \item{STATE}{The U.S. postal code for the state (for U.S. stations
 only).}
 \item{COUNTY}{The name of the county.}
 \item{DAYS}{Number of days over which the maximum accumulated.}
 \item{COUNTY_MAX}{Maximum snow in mm of snowfal accumulated over the number
 of given DAYS.}
 }
}
\usage{
county_maxes
}
\description{
Collected from
https://www.ncdc.noaa.gov/snow-and-ice/snowfall-extremes/. 1, 2, and 3 day
data are from this source. 4,5,...10 day data are calculated by day
* (3day_value / 3).
}
\keyword{datasets}
