% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_all_stations.R
\name{download_all_stations}
\alias{download_all_stations}
\title{Download all .dly weather data from the web.}
\usage{
download_all_stations(directory = ".")
}
\arguments{
\item{directory}{The directory to download the .dly files. Default is the
current working directory.}
}
\value{
Uncompressed ghcnd_all folder and ghcnd-version.txt in the supplied
directory. The source parameter for \code{\link{get_station_data}} and
\code{\link{get_state_data}} can be set to the ghcnd_all folder.
}
\description{
This function is designed to facilitate faster creation of data sets in
the \code{\link{get_station_data}} and \code{\link{get_state_data}}
functions. This is done by downloading all .dly files available, then
using the downloaded folder as the "source" instead of the web source.
}
\seealso{
\code{\link{get_station_data}} - get data by station.

  \code{\link{get_state_data}} - get data by state.
}
