% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dist.R
\name{fit_dist}
\alias{fit_dist}
\title{Fit distributions for snowload yearly maximums.}
\usage{
fit_dist(station_data, id, values, distr, method = "mle", tail = 1, shape, ...)
}
\arguments{
\item{station_data}{A data frame with yearly maximum station snow data (see
\code{\link{yearly_maximums}} to get yearly maximums).}

\item{id}{Column containing station ID's, may include other grouping
variables.}

\item{distr}{A distribution for fitting the data. Must be one of "lnorm" (log
normal), "gamma", "gumbel", or "gev" (generalized extreme value).}

\item{method}{Method "mle" (maximum likelihood) will work for any
distribution. Method "mme" (moment matching) will work with "lnorm" and
"gamma" distributions. Method "lmoments" will only work with "gev"
distribution. Method "regression" will only work with "lnorm" distribution.}

\item{tail}{Proportion of data to use when fitting from 0 to 1. If .33 is
chosen, then only the upper third of the data will be used for fitting.
Only applies to "regression" method.}

\item{shape}{Column containing shape paramter when shape parameter is to be
fixed. Only applies to "gev" distr with "mle" method.}

\item{...}{Other parameters to send to the distribution functions
fitdistrplus::fitdist for "lnorm" and "gamma" distributions or extRemes::fevd
for "gumbel" and "gev" distributions.}

\item{value}{Column containing the observations to maximize}
}
\value{
A data.frame where each row is a maximum for a given id and year:
  \describe{
    \item{}{\emph{id} - Given id column and possibly other grouping columns.}
    \item{}{\emph{DISTR} - Which distribution was fit.}
    \item{}{\emph{NFIT} - Number of values used for fitting. (The number
      of values used to fit the distribution is (1 - PROP_ZERO) * NFIT.)}
    \item{}{\emph{PROP_ZERO} - Proportion of zeros in the values used for
      fitting.}
    \item{}{\emph{PRIORITIZED} - Ratio of prioritized values used for fitting.
     This column will only appear if \emph{station_data} has a logical
     column called PRIORITIZED from \code{\link{yearly_maximums}}.}
    \item{}{\emph{PAR1} - First fitting parameter. For "lnorm": meanlog,
      "gamma": shape, "gumbel": location, and "gev": location.}
    \item{}{\emph{PAR2} - Second fitting parameter. For "lnorm": sdlog,
      "gamma": rate, "gumbel": scale, and "gev": scale.}
    \item{}{\emph{PAR3} - Third fitting parameter. For "gev": shape.}
    \item{}{\emph{EVENT50} - Predicted value for 50 year event.}
  }
}
\description{
This function will group by id variable(s) and return distribution parameters
for each group of values.
}
\examples{
x <- get_station_data(ghcnd_stations$ID[10000],
                      "ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/all/")
xym <- yearly_maximums(x, ID, DATE, VALUE)
fit_dist(xym, ID, MAX, distr = "gev", method = "lmoments")


}
\seealso{
\code{\link{yearly_maximums}} - get maximum values by "water year".
}
