% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_data.R
\name{get_station_data}
\alias{get_station_data}
\alias{get_state_data}
\title{Organize .dly weather data downloaded from the web.}
\usage{
get_station_data(stations, source, elem = c("SNWD", "WESD"), progress = TRUE)

get_state_data(states, source, elem = c("SNWD", "WESD"), progress = TRUE)
}
\arguments{
\item{stations}{A vector of station identification codes. See the internal
data set ghcnd_stations for station id's and descriptions.}

\item{source}{If downloading from the web, source should be
"ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/all/". For loading many
stations/states it is faster to use \code{\link{get_year_data}} (~30GB),
then use the file path to the resulting folder ghcnd_all as the source.}

\item{elem}{The climate variables to extract. Default is snow depth in mm
("SNWD") and water depth equivalent of snowfall in mm ("WESD"). For other
climate variables see section three of
ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt.}

\item{progress}{If true, a text progress bar is printed to the console.
Only relevant when more than one station is requested.}

\item{states}{A vector of states, see the internal data set ghcnd_stations
for valid states and associated stations.}
}
\value{
A data.frame where each row is a measurement of a climate variable
  in elem for a single day and station. Columns are:
  \describe{
    \item{}{\emph{ID} - The station identification code. See the internal
    data set ghcnd_stations for station id's and descriptions.}
    \item{}{\emph{DATE} - The date of the measurement.}
    \item{}{\emph{ELEMENT} - The climate variable measured. See
    ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt, section 3  for
    a description of climate variables.}
    \item{}{\emph{VALUE} - The measurement of the ELEMENT on the given
    DATE.}
    \item{}{\emph{MFLAG} - A measurement flag. See
    ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt, section 3  for
    a description of the flag.}
    \item{}{\emph{QFLAG} - A quality flag. See
    ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt, section 3  for
    a description of the flag.}
    \item{}{\emph{SFLAG} - A source flag. See
    ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt, section 3  for
    a description of the flag.}
  }
}
\description{
These functions are designed to take NOAA weather data organized in .dly
format, extract snow depth, WESD, or other climate information, and return
a neatly organized data frame.
}
\section{Functions}{
\itemize{
\item \code{get_station_data}: Uses station identification codes to gather
data from the proper .dly files.

\item \code{get_state_data}: Wrapper for get_station_data that extracts all
station data for a given state(s) by passing all stations from internal
data source ghcnd_stations with corresponding state(s).
}}

\examples{
get_station_data(ghcnd_stations$ID[10000],
                 "ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/all/")


}
\seealso{
\code{\link{get_year_data}} - get data by year.

  \code{\link{download_all_stations}} - download all available .dly data.
}
