% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_year_data.R
\name{get_year_data}
\alias{get_year_data}
\title{Organize .csv weather data downloaded from the web.}
\usage{
get_year_data(year, source, elem = c("SNWD", "WESD"))
}
\arguments{
\item{year}{A vector of years where data is to be gathered.}

\item{source}{If downloading from the web, source should be
"ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/all/by_year". Data can be
manually downloaded, then use the file path to the folder with
extracted .csv files as the source.}

\item{elem}{The climate variables to extract. Default is snow depth in mm
("SNWD") and water depth equivalent of snowfall in mm ("WESD"). For other
climate variables see section three of
ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt.}
}
\value{
A data.frame where each row is a measurement of a climate variable
  in elem for a single day and station. Columns are:
  \describe{
    \item{}{\emph{ID} - The station identification code. See the internal
    data set ghcnd_stations for station id's and descriptions.}
    \item{}{\emph{DATE} - The date of the measurement.}
    \item{}{\emph{ELEMENT} - The climate variable measured. See
    ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt, section 3  for
    a description of climate variables.}
    \item{}{\emph{VALUE} - The measurement of the ELEMENT on the given
    DATE.}
    \item{}{\emph{MFLAG} - A measurement flag. See
    ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt, section 3  for
    a description of the flag.}
    \item{}{\emph{QFLAG} - A quality flag. See
    ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt, section 3  for
    a description of the flag.}
    \item{}{\emph{SFLAG} - A source flag. See
    ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt, section 3  for
    a description of the flag.}
  }
}
\description{
Takes yearly NOAA weather data organized in .csv format, extract snow depth,
WESD, or other climate information, and return a neatly organized data frame.
}
\examples{
get_year_data(1770,
              "ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/by_year/",
              elem = c("TMAX", "TMIN"))

}
\seealso{
\code{\link{get_station_data}} - get data by station.

  \code{\link{get_state_data}} - get data by state.
}
