% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ghcnd_stations}
\alias{ghcnd_stations}
\title{Global Historical Climatology Network (GHCN-Daily) Stations}
\format{}
\usage{
ghcnd_stations
}
\description{
Collected from
ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/ghcnd-stations.txt.
}
\keyword{datasets}
