% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_maximums.R
\name{monthly_maximums}
\alias{monthly_maximums}
\title{Get the maximum values by month.}
\usage{
monthly_maximums(station_data, id, date, value, value_type)
}
\arguments{
\item{station_data}{A data frame with station snow data.}

\item{id}{Column containing station ID's, may include other grouping
variables.}

\item{date}{Column containing the observation date.}

\item{value}{Column containing the observations to maximize}

\item{value_type}{Optional column containing the value types. If a value_type
column is specified, maximums will be found for each  unique value_type.}
}
\value{
A data.frame where each row is a maximum for a given id and month.
Columns will be the same as station_data input with the addition of:
  \describe{
    \item{}{\emph{MONTH} - Month for a given maximum.}
    \item{}{\emph{YEAR} - The "water years" of the given dates.}
    \item{}{\emph{MAX_N} - Number of values used to find the max.}
  }
}
\description{
This function will group by id variable(s) and return the maximum values
per month. "Water year" is defined as a 12 month period  that begins
October 1 and ends September 30. The year is the calender year in which the
period ends.
}
\examples{
x <- get_station_data(ghcnd_stations$ID[10000],
                      "ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/all/")
monthly_maximums(x, ID, DATE, VALUE)


}
\seealso{
\code{\link{yearly_maximums}} - get maxes by year.
}
