% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outlier_combined}
\alias{outlier_combined}
\title{Removed values flagged by state and county max filters.}
\format{
A data frame with 664,324 observations and 7 variables:

 \describe{
 \item{ID}{The GHCND station identifier.}
 \item{DATE}{Day in yyyy-mm-dd format.}
 \item{ELEMENT}{Value indicating water equivalent of snow on the ground
   (WESD) or snow depth (SNWD).}
 \item{VALUE}{Either WESD in tenths of millimeters or SNWD in millimeters.}
 \item{FLAGGED}{0 - Value not flagged but manually detected during
 outlier search. 1 - Value flagged for exceeding state extreme snow depth.
 2- Value flagged for exceeding county extreme snowfall.}
 \item{OUTLIER}{Manual determination if the value was, in fact, and outlier.
 1 - yes, 0 - no, -1 - values removed for coverage issue (all zero values).}
 \item{TYPE}{Method in which outlier was detected. Options include:
 \itemize{
 \item state_flag: value was flagged for exceeding state snow depth or county snowfall records.
 \item heavy_tail: observations were investigated from a distribution with an unusually large, positive shape parameter.
 \item left_skew: observations were investigated from a distribution with an unusually large, negative shape parameter.
 \item region_shape: observations were investigated from a distribution with a shape parameter notably different than the
 shape parameter predicted via regional smoothing.
 \item yr50: observations were investigated because the resulting 50 year event was very different than the maximum max observation.
 \item state_scatter: observations were identified as having 50 year events significantly different from other observations in the same state.
 \item heavy_tail_2: observations with large, positive shape parameters as well as a maximum max 4 times larger than the estimated 50 year event.
 \item OR_WA_check: observations from a region-level investigation of observations in coastal Oregon and Washington.
 }}
 }
}
\usage{
outlier_combined
}
\description{
Collected from state and county max data available at
https://www.ncdc.noaa.gov/snow-and-ice/snowfall-extremes/ and
https://www.ncdc.noaa.gov/extremes/scec/. All records are given
for snow depth and snowfall. WESD outliers are detected by assuming
a specific gravity of 0.5 (which is intentionally conservative).
}
\keyword{datasets}
