% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outlier_combined_alt}
\alias{outlier_combined_alt}
\title{Alternate outliers as verified through secondary manual checks.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 696797 rows and 5 columns.
}
\usage{
outlier_combined_alt
}
\description{
\describe{
 \item{ID}{The GHCND station identifier.}
 \item{DATE}{Day in yyyy-mm-dd format.}
 \item{ELEMENT}{Value indicating water equivalent of snow on the ground
   (WESD) or snow depth (SNWD).}
 \item{VALUE}{Either WESD in tenths of millimeters or SNWD in millimeters.}
 \item{OUTLIER}{Manual determination if the value was, in fact, and outlier.
 1 - yes, 0 - no, -1 - values removed for coverage issue (all zero values).}
 }
}
\keyword{datasets}
