% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outlier_state_county}
\alias{outlier_state_county}
\title{Removed values flagged by state and county max filters.}
\format{
A data frame with 303,121 observations and 6 variables:

 \describe{
 \item{ID}{The GHCND station identifier.}
 \item{DATE}{Day in yyyy-mm-dd format.}
 \item{ELEMENT}{Value indicating water equivalent of snow on the ground
   (WESD) or snow depth (SNWD).}
 \item{VALUE}{Either WESD in tenths of millimeters or SNWD in millimeters.}
 \item{FLAGGED}{0 - Value not flagged but manually detected during
 outlier search. 1 - Value flagged for exceeding state extreme snow depth.
 2- Value flagged for exceeding county extreme snowfall.}
 }
}
\usage{
outlier_state_county
}
\description{
Collected from state and county max data available at
https://www.ncdc.noaa.gov/snow-and-ice/snowfall-extremes/ and
https://www.ncdc.noaa.gov/extremes/scec/. All records are given
for snow depth and snowfall. WESD outliers are detected by assuming
a specific gravity of 0.5 (which is intentionally conservative).
}
\keyword{datasets}
