% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev_quantiles_vec.R
\name{pgev}
\alias{pgev}
\title{GEV Distribution Quantile Estimation}
\usage{
pgev(x, loc = 0, scale = 1, shape = 0)
}
\arguments{
\item{x}{A realization from the distribution.}

\item{loc}{The location parameter, may be a vector.}

\item{scale}{The scale parameter, may be a vector.}

\item{shape}{The shape parameter, may be a vector.}
}
\value{
A vector of quantiles.
}
\description{
Calculates the value obtained from the Generalized Extreme Value
  Distribution for a specific quantile. Vectorized for speed.
}
