% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev_quantiles_vec.R
\name{qgev}
\alias{qgev}
\title{GEV Distribution Value Estimation From Quantile}
\usage{
qgev(p, loc = 0, scale = 1, shape = 0, tol = 1e-06)
}
\arguments{
\item{p}{A quantile between 0 and 1, may be a vector.}

\item{loc}{The location parameter, may be a vector.}

\item{scale}{The scale parameter, may be a vector.}

\item{shape}{The shape parameter, may be a vector.}
}
\value{
A vector of values from the distribution based on the
  the provided quantiles.
}
\description{
Calculates the value obtained from the Generalized Extreme Value
  Distribution for a specific quantile. Vectorized for speed.
}
