% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_maximums.R
\name{yearly_maximums}
\alias{yearly_maximums}
\title{Get the maximum values by "water year".}
\usage{
yearly_maximums(station_data, id, date, value)
}
\arguments{
\item{station_data}{A data frame with station snow data.}

\item{id}{Column containing station ID's, may include other grouping
variables.}

\item{date}{Column containing the observation date.}

\item{value}{Column containing the observations to maximize}
}
\value{
A data.frame where each row is a maximum for a given id and year:
  \describe{
    \item{}{\emph{id} - Given id column and possibly other grouping columns.}
    \item{}{\emph{YEAR} - The "water years" of the given dates.}
    \item{}{\emph{MAX} - The maximum of the given values.}
  }
}
\description{
This function will group by id variable(s) and return the maximum values
per "water year". "Water year" is defined as a 12 month period  that begins
October 1 and ends September 30. The year is the calender year in which the
period ends.
}
\examples{
x <- get_station_data(ghcnd_stations$ID[10000],
                      "ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/all/")
yearly_maximums(x, ID, DATE, VALUE)


}
