% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_maximums.R
\name{yearly_maximums}
\alias{yearly_maximums}
\title{Get the maximum values by "water year".}
\usage{
yearly_maximums(
  station_data,
  id,
  date,
  value,
  ratio = NULL,
  value_type = NULL,
  prioritize = NULL,
  prioritize_threshold = list(n = 10, months = 3)
)
}
\arguments{
\item{station_data}{A data frame with station snow data.}

\item{id}{Column containing station ID's, may include other grouping
variables.}

\item{date}{Column containing the observation date.}

\item{value}{Column containing the observations to maximize}

\item{value_type}{Optional column containing the value types.}

\item{prioritize}{Optional value_type(s) to prioritize when selecting.}

\item{prioritize_threshold}{Optional named list with requirements to use
the prioritized value containing \emph{n} - the minimum number of
observations and \emph{months} - the minimum number of months that must have
a value.}
}
\value{
A data.frame where each row is a maximum for a given id and year:
  \describe{
    \item{}{\emph{id} - Given id column and possibly other grouping columns.}
    \item{}{\emph{YEAR} - The "water years" of the given dates.}
    \item{}{\emph{MAX} - The maximum of the given values.}
    \item{}{\emph{MAX_N} - Number of values used to find the max.}
    \item{}{\emph{MAX_MONTHS} - Number of unique months with values.}
    \item{}{\emph{MAX_MONTH} - Month in which the maximum was found.}
    \item{}{\emph{PRIORITIZED} - Only applies when \emph{prioritize} is not
    null. True when max value is a prioritized value, false otherwise.}
  }
}
\description{
This function will group by id variable(s) and return the maximum values
per "water year". "Water year" is defined as a 12 month period  that begins
October 1 and ends September 30. The year is the calender year in which the
period ends.
}
\details{
If a certain value type is to be prioritized, the value_type column should
be specified along with the specific value_type(s) to prioritize (e.g.
value_type = ELEMENT, prioritize = "WESD"). Values that have priority will
be chosen over non-prioritized values for a given year (e.g. if the maximum
value is for the year is 300 for type "SNWD" but the prioritized "WESD"
has max value 250, 250 will be chosen as the maximum).
}
\examples{
x <- get_station_data(ghcnd_stations$ID[10000],
                      "ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/all/")
yearly_maximums(x, ID, DATE, VALUE)


}
